Extract share logic into classes, put share URL into challenge attempt response
